d-bias/backend/requirements.txt
flask
pandas
numpy
scipy
scikit-learn
fairlearn
plotly
google-generativeai
supabase
python-dotenv